fix wallet corruption on rescan and zapwallettxes (#189)
* reenable zapwallettxes.py test

* add additional flush in shutdown, clean up threadgroup with join_all

* compare min version to wallet version not client version

* remove account tests (dont support accounts), enable rescan,reindex,zap,salvage tests

* make cs_utxo a shared mutex and rework the coin logic accordingly

* rework validation code, fix rescan failures caused by disconnect failure

* fix formatting

* change getchainips to explicitly reconnect the network

+++ src/rpc/rpcblockchain.cpp
@@ -627,7 +627,7 @@ UniValue gettxout(const UniValue &params, bool fHelp)
     if ((unsigned int)coin.nHeight == MEMPOOL_HEIGHT)
         ret.push_back(Pair("confirmations", 0));
     else
-        ret.push_back(Pair("confirmations", pindex->nHeight - coin.nHeight + 1));
+        ret.push_back(Pair("confirmations", (int64_t)(pindex->nHeight - coin.nHeight + 1)));
     ret.push_back(Pair("value", ValueFromAmount(coin.out.nValue)));
     UniValue o(UniValue::VOBJ);
     ScriptPubKeyToJSON(coin.out.scriptPubKey, o, true);
